+++ transport/http/handler_test.go
@@ -108,6 +108,7 @@ func TestHandlerHeaders(t *testing.T) {
 	tests := []struct {
 		giveEncoding string
 		giveHeaders  http.Header
+		grabHeaders  map[string]struct{}
 
 		wantTTL     time.Duration
 		wantHeaders map[string]string
@@ -117,10 +118,13 @@ func TestHandlerHeaders(t *testing.T) {
 			giveHeaders: http.Header{
 				TTLMSHeader:      {"1000"},
 				"Rpc-Header-Foo": {"bar"},
+				"X-Baz":          {"bat"},
 			},
-			wantTTL: time.Second,
+			grabHeaders: map[string]struct{}{"x-baz": {}},
+			wantTTL:     time.Second,
 			wantHeaders: map[string]string{
-				"foo": "bar",
+				"foo":   "bar",
+				"x-baz": "bat",
 			},
 		},
 		{
@@ -160,7 +164,8 @@ func TestHandlerHeaders(t *testing.T) {
 			WithProcedure("hello"),
 		).Return(spec, nil)
 
-		httpHandler := handler{router: router, tracer: nil}
+
+		httpHandler := handler{router: router, tracer: nil, grabHeaders: tt.grabHeaders}
 
 		rpcHandler.EXPECT().Handle(
 			transporttest.NewContextMatcher(t,