+++ CHANGELOG.md
@@ -4,6 +4,8 @@ Releases
 v1.17.0-dev (unreleased)
 --------------------
 
+-   transport/http: Add GrabHeaders option to propagate specific
+    headers starting with x- to handlers.
 -   Export no-op backoff strategy in api/backoff.