+++ CHANGELOG.md
@@ -4,7 +4,9 @@ Releases
 v1.17.0-dev (unreleased)
 --------------------
 
--   No changes yet.
+-   transport/http: Add GrabHeaders option to propagate specific
+    headers starting with x- to handlers.
+-   Export no-op backoff strategy in api/backoff.
 
 
 v1.16.0 (2017-09-18)